+++ src/gicv3/cpu_interface.rs
@@ -5,12 +5,13 @@ use crate::{
     IntId,
     gicv3::GicError,
     sysreg::{
-        IccIgrpen1El3, IccIgrpenEl1, IccSreEl1, IccSreEl23, Sgir, read_icc_hppir0_el1,
-        read_icc_hppir1_el1, read_icc_iar0_el1, read_icc_iar1_el1, read_icc_igrpen1_el3,
-        read_icc_pmr_el1, read_icc_sre_el1, read_icc_sre_el2, read_icc_sre_el3,
-        write_icc_asgi1r_el1, write_icc_eoir0_el1, write_icc_eoir1_el1, write_icc_igrpen0_el1,
-        write_icc_igrpen1_el1, write_icc_igrpen1_el3, write_icc_pmr_el1, write_icc_sgi0r_el1,
-        write_icc_sgi1r_el1, write_icc_sre_el1, write_icc_sre_el2, write_icc_sre_el3,
+        IccIgrpen0El1, IccIgrpen1El1, IccIgrpen1El3, IccPmrEl1, IccSreEl1, IccSreEl2, IccSreEl3,
+        Sgir, read_icc_hppir0_el1, read_icc_hppir1_el1, read_icc_iar0_el1, read_icc_iar1_el1,
+        read_icc_igrpen1_el3, read_icc_pmr_el1, read_icc_sre_el1, read_icc_sre_el2,
+        read_icc_sre_el3, write_icc_asgi1r_el1, write_icc_eoir0_el1, write_icc_eoir1_el1,
+        write_icc_igrpen0_el1, write_icc_igrpen1_el1, write_icc_igrpen1_el3, write_icc_pmr_el1,
+        write_icc_sgi0r_el1, write_icc_sgi1r_el1, write_icc_sre_el1, write_icc_sre_el2,
+        write_icc_sre_el3,
     },
 };
 #[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
@@ -26,32 +27,32 @@ impl GicCpuInterface {
     /// Enables or disables group 0 interrupts.
     pub fn enable_group0(enable: bool) {
         write_icc_igrpen0_el1(if enable {
-            IccIgrpenEl1::EN
+            IccIgrpen0El1::ENABLE
         } else {
-            IccIgrpenEl1::empty()
+            IccIgrpen0El1::empty()
         });
     }
 
     /// Enables or disables group 1 interrupts for the current security state.
     pub fn enable_group1(enable: bool) {
         write_icc_igrpen1_el1(if enable {
-            IccIgrpenEl1::EN
+            IccIgrpen1El1::ENABLE
         } else {
-            IccIgrpenEl1::empty()
+            IccIgrpen1El1::empty()
         });
     }
 
     /// Enables or disables group 1 secure interrupts.
     pub fn enable_group1_secure(enable: bool) {
         let mut value = read_icc_igrpen1_el3();
-        value.set(IccIgrpen1El3::GRP1S, enable);
+        value.set(IccIgrpen1El3::ENABLEGRP1S, enable);
         write_icc_igrpen1_el3(value);
     }
 
     /// Enables or disables group 1 non-secure interrupts.
     pub fn enable_group1_non_secure(enable: bool) {
         let mut value = read_icc_igrpen1_el3();
-        value.set(IccIgrpen1El3::GRP1NS, enable);
+        value.set(IccIgrpen1El3::ENABLEGRP1NS, enable);
         write_icc_igrpen1_el3(value);
     }
 
@@ -60,8 +61,8 @@ impl GicCpuInterface {
     /// Returns `None` if there is no pending interrupt of sufficient priority.
     pub fn get_pending_interrupt(group: InterruptGroup) -> Option<IntId> {
         let intid = IntId(match group {
-            InterruptGroup::Group0 => read_icc_hppir0_el1(),
-            InterruptGroup::Group1 => read_icc_hppir1_el1(),
+            InterruptGroup::Group0 => read_icc_hppir0_el1().intid(),
+            InterruptGroup::Group1 => read_icc_hppir1_el1().intid(),
         });
 
         if intid == IntId::SPECIAL_NONE {
@@ -76,8 +77,8 @@ impl GicCpuInterface {
     /// Returns `None` if there is no pending interrupt of sufficient priority.
     pub fn get_and_acknowledge_interrupt(group: InterruptGroup) -> Option<IntId> {
         let intid = IntId(match group {
-            InterruptGroup::Group0 => read_icc_iar0_el1(),
-            InterruptGroup::Group1 => read_icc_iar1_el1(),
+            InterruptGroup::Group0 => read_icc_iar0_el1().intid(),
+            InterruptGroup::Group1 => read_icc_iar1_el1().intid(),
         });
 
         if intid == IntId::SPECIAL_NONE {
@@ -91,8 +92,8 @@ impl GicCpuInterface {
     /// This drops the interrupt priority and deactivates the interrupt.
     pub fn end_interrupt(intid: IntId, group: InterruptGroup) {
         match group {
-            InterruptGroup::Group0 => write_icc_eoir0_el1(intid.0),
-            InterruptGroup::Group1 => write_icc_eoir1_el1(intid.0),
+            InterruptGroup::Group0 => write_icc_eoir0_el1(intid.into()),
+            InterruptGroup::Group1 => write_icc_eoir1_el1(intid.into()),
         }
     }
 
@@ -117,9 +118,9 @@ impl GicCpuInterface {
         };
 
         match group {
-            SgiTargetGroup::Group0 => write_icc_sgi0r_el1(sgir),
-            SgiTargetGroup::CurrentGroup1 => write_icc_sgi1r_el1(sgir),
-            SgiTargetGroup::OtherGroup1 => write_icc_asgi1r_el1(sgir),
+            SgiTargetGroup::Group0 => write_icc_sgi0r_el1(sgir.into()),
+            SgiTargetGroup::CurrentGroup1 => write_icc_sgi1r_el1(sgir.into()),
+            SgiTargetGroup::OtherGroup1 => write_icc_asgi1r_el1(sgir.into()),
         }
 
         Ok(())
@@ -129,71 +130,90 @@ impl GicCpuInterface {
     ///
     /// Only interrupts with a higher priority (numerically lower) will be signalled.
     pub fn set_priority_mask(min_priority: u8) {
-        write_icc_pmr_el1(min_priority.into());
+        write_icc_pmr_el1(IccPmrEl1::empty().with_priority(min_priority));
     }
 
     /// Gets the priority mask for the current CPU core.
     pub fn get_priority_mask() -> u8 {
-        read_icc_pmr_el1() as u8
+        read_icc_pmr_el1().priority()
     }
 
-    /// Enables or disables system register interface for the current Security state.
-    pub fn enable_system_register_el1(enable: bool) {
-        let mut value = read_icc_sre_el1();
-        value.set(IccSreEl1::SRE, enable);
-        write_icc_sre_el1(value);
+    /// Enables the EL1 system register interface for the current security state.
+    pub fn enable_system_register_el1() {
+        // SAFETY: Changing the SRE bit from 0 to 1 is safe.
+        unsafe {
+            write_icc_sre_el1(read_icc_sre_el1() | IccSreEl1::SRE);
+        }
     }
 
-    /// Enables or disables the system register interface to the ICH_* registers and the EL1 and
-    /// EL2 ICC_* registers for EL2. The `enable_lower` parameter controls the lower exception level
-    /// access to ICC_SRE_EL1.
-    pub fn enable_system_register_el2(enable: bool, enable_lower: bool) {
+    /// Enables the system register interface to the `ICH_*` registers and the EL1 and EL2 `ICC_*`
+    /// registers for EL2. The `enable_lower` parameter controls the lower exception level access to
+    /// `ICC_SRE_EL1`.
+    pub fn enable_system_register_el2(enable_lower: bool) {
         let mut value = read_icc_sre_el2();
-        value.set(IccSreEl23::SRE, enable);
-        value.set(IccSreEl23::ENABLE, enable_lower);
-        write_icc_sre_el2(value);
+        value.set(IccSreEl2::SRE, true);
+        value.set(IccSreEl2::ENABLE, enable_lower);
+        // SAFETY: Changing the SRE bit from 0 to 1 is safe, and changing the Enable bit either way
+        // is safe.
+        unsafe {
+            write_icc_sre_el2(value);
+        }
     }
 
-    /// Enables or disabled the system register interface to the ICH_* registers and the EL1, EL2,
-    /// and EL3 ICC_* registers for EL3. The `enable_lower` parameter controls the lower exception
-    /// level access to ICC_SRE_EL1 and ICC_SRE_EL2.
-    pub fn enable_system_register_el3(enable: bool, enable_lower: bool) {
+    /// Enables the system register interface to the `ICH_*` registers and the EL1, EL2, and EL3
+    /// `ICC_*` registers for EL3. The `enable_lower` parameter controls the lower exception level
+    /// access to `ICC_SRE_EL1` and `ICC_SRE_EL2`.
+    pub fn enable_system_register_el3(enable_lower: bool) {
         let mut value = read_icc_sre_el3();
-        value.set(IccSreEl23::SRE, enable);
-        value.set(IccSreEl23::ENABLE, enable_lower);
-        write_icc_sre_el3(value);
+        value.set(IccSreEl3::SRE, true);
+        value.set(IccSreEl3::ENABLE, enable_lower);
+        // SAFETY: Changing the SRE bit from 0 to 1 is safe, and changing the Enable bit either way
+        // is safe.
+        unsafe {
+            write_icc_sre_el3(value);
+        }
     }
 
     /// Disables IRQ and FIQ bypass for EL1.
     pub fn disable_legacy_interrupt_bypass_el1(disable: bool) {
         let mut value = read_icc_sre_el1();
         value.set(IccSreEl1::DFB | IccSreEl1::DIB, disable);
-        write_icc_sre_el1(value);
+        // SAFETY: We don't change the SRE bit.
+        unsafe {
+            write_icc_sre_el1(value);
+        }
     }
 
     /// Disables IRQ and FIQ bypass for EL2.
     pub fn disable_legacy_interrupt_bypass_el2(disable: bool) {
         let mut value = read_icc_sre_el2();
-        value.set(IccSreEl23::DFB | IccSreEl23::DIB, disable);
-        write_icc_sre_el2(value);
+        value.set(IccSreEl2::DFB | IccSreEl2::DIB, disable);
+        // SAFETY: We don't change the SRE bit.
+        unsafe {
+            write_icc_sre_el2(value);
+        }
     }
 
     /// Disables IRQ and FIQ bypass for EL3.
     pub fn disable_legacy_interrupt_bypass_el3(disable: bool) {
         let mut value = read_icc_sre_el3();
-        value.set(IccSreEl23::DFB | IccSreEl23::DIB, disable);
-        write_icc_sre_el3(value);
+        value.set(IccSreEl3::DFB | IccSreEl3::DIB, disable);
+        // SAFETY: We don't change the SRE bit.
+        unsafe {
+            write_icc_sre_el3(value);
+        }
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::sysreg::fake::SYSREGS;
-
     use super::*;
+    use arm_sysregs::{
+        IccEoir0El1, IccEoir1El1, IccHppir0El1, IccHppir1El1, IccIar0El1, IccIar1El1, fake::SYSREGS,
+    };
 
     fn clear_regs() {
-        SYSREGS.lock().unwrap().clear();
+        SYSREGS.lock().unwrap().reset();
     }
 
     macro_rules! read_reg {
@@ -251,19 +271,19 @@ mod tests {
     fn get_pending_interrupt() {
         clear_regs();
 
-        write_reg!(icc_hppir0_el1, 1);
+        write_reg!(icc_hppir0_el1, IccHppir0El1::from_bits_retain(1));
         assert_eq!(
             Some(IntId::sgi(1)),
             GicCpuInterface::get_pending_interrupt(InterruptGroup::Group0)
         );
 
-        write_reg!(icc_hppir1_el1, 16);
+        write_reg!(icc_hppir1_el1, IccHppir1El1::from_bits_retain(16));
         assert_eq!(
             Some(IntId::ppi(0)),
             GicCpuInterface::get_pending_interrupt(InterruptGroup::Group1)
         );
 
-        write_reg!(icc_hppir0_el1, 1023);
+        write_reg!(icc_hppir0_el1, IccHppir0El1::from_bits_retain(1023));
         assert_eq!(
             None,
             GicCpuInterface::get_pending_interrupt(InterruptGroup::Group0)
@@ -274,19 +294,19 @@ mod tests {
     fn get_and_acknowledge_interrupt() {
         clear_regs();
 
-        write_reg!(icc_iar0_el1, 32);
+        write_reg!(icc_iar0_el1, IccIar0El1::from_bits_retain(32));
         assert_eq!(
             Some(IntId::spi(0)),
             GicCpuInterface::get_and_acknowledge_interrupt(InterruptGroup::Group0)
         );
 
-        write_reg!(icc_iar1_el1, 64);
+        write_reg!(icc_iar1_el1, IccIar1El1::from_bits_retain(64));
         assert_eq!(
             Some(IntId::spi(32)),
             GicCpuInterface::get_and_acknowledge_interrupt(InterruptGroup::Group1)
         );
 
-        write_reg!(icc_iar0_el1, 1023);
+        write_reg!(icc_iar0_el1, IccIar0El1::from_bits_retain(1023));
         assert_eq!(
             None,
             GicCpuInterface::get_and_acknowledge_interrupt(InterruptGroup::Group0)
@@ -298,10 +318,10 @@ mod tests {
         clear_regs();
 
         GicCpuInterface::end_interrupt(IntId::sgi(15), InterruptGroup::Group0);
-        assert_eq!(15, read_reg!(icc_eoir0_el1));
+        assert_eq!(IccEoir0El1::from_bits_retain(15), read_reg!(icc_eoir0_el1));
 
         GicCpuInterface::end_interrupt(IntId::ppi(15), InterruptGroup::Group1);
-        assert_eq!(31, read_reg!(icc_eoir1_el1));
+        assert_eq!(IccEoir1El1::from_bits_retain(31), read_reg!(icc_eoir1_el1));
     }
 
     #[test]
@@ -353,39 +373,30 @@ mod tests {
     fn enable_system_register_el1() {
         clear_regs();
 
-        GicCpuInterface::enable_system_register_el1(true);
+        GicCpuInterface::enable_system_register_el1();
         assert_eq!(0x0000_0001, read_reg!(icc_sre_el1).bits());
-
-        GicCpuInterface::enable_system_register_el1(false);
-        assert_eq!(0x0000_0000, read_reg!(icc_sre_el1).bits());
     }
 
     #[test]
     fn enable_system_register_el2() {
         clear_regs();
 
-        GicCpuInterface::enable_system_register_el2(true, false);
+        GicCpuInterface::enable_system_register_el2(false);
         assert_eq!(0x0000_0001, read_reg!(icc_sre_el2).bits());
 
-        GicCpuInterface::enable_system_register_el2(true, true);
+        GicCpuInterface::enable_system_register_el2(true);
         assert_eq!(0x0000_0009, read_reg!(icc_sre_el2).bits());
-
-        GicCpuInterface::enable_system_register_el2(false, false);
-        assert_eq!(0x0000_0000, read_reg!(icc_sre_el2).bits());
     }
 
     #[test]
     fn enable_system_register_el3() {
         clear_regs();
 
-        GicCpuInterface::enable_system_register_el3(true, false);
+        GicCpuInterface::enable_system_register_el3(false);
         assert_eq!(0x0000_0001, read_reg!(icc_sre_el3).bits());
 
-        GicCpuInterface::enable_system_register_el3(true, true);
+        GicCpuInterface::enable_system_register_el3(true);
         assert_eq!(0x0000_0009, read_reg!(icc_sre_el3).bits());
-
-        GicCpuInterface::enable_system_register_el3(false, false);
-        assert_eq!(0x0000_0000, read_reg!(icc_sre_el3).bits());
     }
 
     #[test]